Remove RPi exp from auto-discovery support
Even though auto-discovery will still be attempted, it doesn't work over
the RPi cable since the UART metadata is different from the USB serial
metadata and hence we cannot distinguish a Marty from another UART
device. So the filtering used for USB devices will discard the UART
ones. This should not be a concern since /dev/ttyAMA0 should always be
the same on RPi and even if not, it is easy to figure out what it
changed to.